requirements.txt
dnspython
requests
colorama
psutil
playwright
